assets/python/2025-08-16-rags/pyproject.toml
[project]
name = "2025-08-16-rags"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.5",
    "pandas>=2.3.2",
    "plotly>=6.3.0",
    "pyyaml>=6.0.2",
    "sentence-transformers>=5.1.0",
    "umap-learn>=0.5.9.post2",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
]
